requirements-dev.txt
# Build requirements, matching pyproject.toml[build-system][requires]
Cython
oldest-supported-numpy
scipy
setuptools
setuptools_scm[toml]>=6.0
wheel
# Development dependencies
furo>=2021.11.16
ipython  # for pygments syntax highlighting
pytest-cov
sphinx
nbsphinx-link
docutils<0.21 # workaround for https://github.com/vidartf/nbsphinx-link/issues/22